refactor: expand ability of parser
