+++ src/main/java/run/halo/editor/hyperlink/handler/HyperLinkParser.java
@@ -1,11 +1,13 @@
 package run.halo.editor.hyperlink.handler;
 
+import reactor.core.publisher.Mono;
 import run.halo.editor.hyperlink.dto.HyperLinkBaseDTO;
+import java.net.URI;
 
 /**
  * @author LIlGG
  */
 public interface HyperLinkParser<T extends HyperLinkBaseDTO> {
 
-    T parse(String htmlContent);
+    Mono<T> parse(URI linkURI);
 }
